paper-mario-backend/requirements.txt
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
